requirements.txt
flask
facenet-pytorch
torch
torchvision
pillow
numpy
flask-cors